Cleanup iouring #2
